Migrate headings
