+++ frontend/packages/console-dynamic-plugin-sdk/src/app/components/utils/resource-status.tsx
@@ -23,7 +23,7 @@ type ResourceStatusProps = {
  * )
  * ```
  */
-const ResourceStatus: React.FC<ResourceStatusProps> = ({
+export const ResourceStatus: React.FC<ResourceStatusProps> = ({
   additionalClassNames,
   badgeAlt,
   children,
